refactor: breaking up components and some styling
