+++ src/components/customMultiCheckboxFacet/styles.module.scss
@@ -19,10 +19,10 @@
 }
 
 .currentNavigatedLabel {
-  background-color: #a2d2ff;
+  background-color: var(--secondary-blue);
 
   &.checked {
-    background-color: #a2d2ff;
+    background-color: var(--secondary-blue);
     color: #4f4f4f;
     span {
       color: #4f4f4f;
@@ -31,14 +31,14 @@
 }
 
 .checked {
-  background-color: #3a86ff;
+  background-color: var(--primary-blue);
   color: white;
   span {
     color: white;
   }
 
   &:only-child {
-    background-color: #3a86ff;
+    background-color: var(--primary-blue);
     color: white;
     span {
       color: white;
